feat(logging): enable logging default for serve cmd, set env prefix
